Switch default to '%matplotlib inline' for JupyterHub
JupyterHub won't run Javascript, so we can't use the nbAgg interactive backend
for the default.

However using an interactive backend can be quite useful to demonstrate
looking at x, y, RA, Dec, and pixel values so I kept the instructions about
how to swtich and the frame/figure number specification, which doesn't matter
for non-interactive but is helpful for interactive.