Simplify to just 'convert_cat_to_parquet'
Merge 'convert_all_to_parquet' and 'convert_cat_to_parquet'.
Formerly, 'convert_all_*' was reponsible for loading the catalog
and 'convert_cat_*' took a loaded catalog instance.  As this script
is meant in part to be a pedagogical example as well as a simple tool,
the simplification clarifies the key parts.